Seperate print logic into seperate file
